fixing functional bugs

+++ src/main/java/com/gojek/locator/model/GetDriverResponse.java
@@ -15,7 +15,7 @@ public class GetDriverResponse {
 	public GetDriverResponse() {
 		
 	}
-	public GetDriverResponse(int driverId, Float latitude, Float longitude, Integer distance) {
+	public GetDriverResponse(int driverId, Double latitude, Double longitude, Integer distance) {
 		super();
 		this.driverId = driverId;
 		this.latitude = latitude;
@@ -27,10 +27,10 @@ public class GetDriverResponse {
 	private int driverId;
 	
 	@JsonProperty("latitude")
-	private Float latitude;
+	private Double latitude;
 	
 	@JsonProperty("longitude")
-	private Float longitude;
+	private Double longitude;
 	
 	@JsonProperty("distance")
 	private Integer distance;
@@ -43,19 +43,19 @@ public class GetDriverResponse {
 		this.driverId = driverId;
 	}
 
-	public Float getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Float latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public Float getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Float longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
